requirements.txt
kiteconnect==5.0.0
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.1.0
streamlit>=1.28.0
plotly>=5.17.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.3.0
flask-session>=0.5.0
redis>=5.0.0
gunicorn>=21.2.0
azure-storage-blob>=12.19.0
